requirements.txt
wheel>=0.41.0
attrs>=22.1.0
dynaconf>=3.1.7
keyboard>=0.13.5
tabulate>=0.9.0
pytest>=6.2.5
pytest-cov>=3.0.0